Updated to have pyperry put `default_params` into the query string
Currently the adapter only put default params in the body of POST/PUT/DELETE
requests. This changes it to put the default params in the query string of all
requests. This is to conform how default params are handled regardless of the
request type and to also follow the ruby version of perry.